+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve dapp responsiviness #190 #356
 - Opening dialog to mark praise as duplicate should place focus on input #80
+- Prevent unncessary import of mongoose package by frontend #466
 
 ## [0.8.0] - 2022-06-06